requirements.txt

asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.10.0
psycopg2==2.8.6
pytz==2020.4
sqlparse==0.4.1